feat: adding icon to list
