+++ frontend/src/Controllers/Campaign.tsx
@@ -2,19 +2,13 @@ import { useCallback, useEffect, useState, type FC } from 'react'
 import {
   Box,
   Button,
-  Collapse,
   Divider,
   List,
-  ListItemButton,
-  ListItemIcon,
-  ListItemText,
   Stack,
-  type SxProps,
-  type Theme,
   Typography
 } from '@mui/material'
 import { useTranslation } from 'react-i18next'
-import { parseErrorMessage, shrinkText } from '../Utils/f'
+import { parseErrorMessage, shrinkText, schema } from '../Utils/f'
 import {
   clearMasterState,
   fetchACampaign,
@@ -31,17 +25,13 @@ import { faMapLocationDot, faPeopleGroup } from '@fortawesome/free-solid-svg-ico
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { useNavigate } from 'react-router-dom'
-import { ExpandLess, ExpandMore, InboxOutlined, GroupAdd } from '@mui/icons-material'
+import { GroupAdd } from '@mui/icons-material'
 import { type SubmitHandler, useForm } from 'react-hook-form'
+import PointOfInterest from './PointOfInterest'
 import Loader from '../Components/Loader'
 import TextAreaDialog from '../Components/TextAreaDialog'
 import ErrorComponent from '../Components/Error'
 import Text from '../Components/Text'
-import * as yup from 'yup'
-
-const schema = yup.object().shape({
-  text: yup.string()
-})
 
 interface CampaignProps {
   activeCampaign: number
@@ -286,95 +276,3 @@ const Campaign: FC<CampaignProps> = ({ activeCampaign }) => {
 }
 
 export default Campaign
-
-interface PointOfInterestProps {
-  point: number
-  points: Record<number, PlaceOfInterestPoint>
-  places: Record<PlacesOfInterestType, number[]>
-  style?: SxProps<Theme>
-}
-
-const PointOfInterest: FC<PointOfInterestProps> = ({ point, points, places, style }) => {
-  const { t } = useTranslation()
-
-  const {
-    handleSubmit,
-    control,
-    setValue,
-    reset,
-    setError,
-    formState: { errors }
-  } = useForm({
-    resolver: yupResolver(schema),
-    defaultValues: {
-      text: ''
-    }
-  })
-
-  const [open, setOpen] = useState<boolean>(false)
-  const [openDescriptionEdit, setOpenDescriptionEdit] = useState<boolean>(false)
-
-  const handleClick = useCallback(() => {
-    setOpen(!open)
-  }, [open])
-  const closeDescription = useCallback(() => {
-    setOpenDescriptionEdit(false)
-  }, [])
-  const openDescription = useCallback(() => {
-    setValue('text', points[point].description)
-    setOpenDescriptionEdit(true)
-  }, [point, points, setValue])
-  const cancelDescription = useCallback(() => {
-    reset()
-  }, [reset])
-
-  const onSubmit: SubmitHandler<FormDataText> = useCallback(async (data) => {
-    try {
-      const description = data.text ?? ''
-      console.log('update poi description', description)
-      setOpenDescriptionEdit(false)
-    } catch (e) {
-      const msg = parseErrorMessage((e))
-      setError('text', { type: 'custom', message: msg ?? 'validation failed' }, { shouldFocus: true })
-    }
-  }, [setError])
-
-  const pointOfInterest = points[point]
-
-  return <Box display='flex' flexDirection='column' sx={{ ...style }}>
-    <TextAreaDialog
-      open={openDescriptionEdit}
-      control={control}
-      errors={errors}
-      onSubmit={onSubmit}
-      handleSubmit={handleSubmit}
-      close={closeDescription}
-      cancel={cancelDescription}
-      body={t('activeCampaign.poiBody')}
-      title={t('activeCampaign.poiTitle')}
-    />
-    <ListItemButton onClick={handleClick}>
-      <ListItemIcon>
-        <InboxOutlined />
-      </ListItemIcon>
-      <ListItemText primary={pointOfInterest.name} />
-      {open ? <ExpandLess /> : <ExpandMore />}
-    </ListItemButton>
-    <Text
-      open={openDescription}
-      chunked={pointOfInterest.description}
-      text={pointOfInterest.description}
-      emptyText={t('activeCampaign.POInoDescription')}
-      button={t('activeCampaign.descriptionButton')}
-      showMore={t('activeCampaign.editDescription')}
-    />
-    <Divider />
-    {
-      pointOfInterest.children.length !== 0 && pointOfInterest.children.map(child =>
-        <Collapse in={open} timeout="auto" unmountOnExit key={child}>
-          <PointOfInterest point={child} points={points} places={places} style={{ pl: 2 }} />
-        </Collapse>
-      )
-    }
-  </Box>
-}
