+++ frontend/src/Utils/f.ts
@@ -1,5 +1,6 @@
 import { usernameIsValid } from '../Api/user'
 import { authorization } from './config'
+import * as yup from 'yup'
 
 export const apply = <T>(x: T, f: (x: T) => T): T => f(x)
 
@@ -92,3 +93,7 @@ export const formatPOI = (poi: PlacesOfInterestDTO): PlacesOfInterest => {
     roots
   } satisfies PlacesOfInterest
 }
+
+export const schema = yup.object().shape({
+  text: yup.string()
+})
